+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Config to use sessionDurationSeconds and sessionRefreshThreshold instead of keepAliveSeconds. 
+  KeepAliveSeconds is now calculated from these two parameters and sessionDurationSeconds is used when starting a 
+  session.
+- Changed KeepAliveSession from long to int.
+- Changed startWidgetSession to take a nullable Integer to set specific Expiry. This should not be lower than 600 
+  due to Kaltura caching of responses.  
 
 ### Added
+- Added getSessionInfo that logs sessionInfo. Only used for testing.
 
 ### Changed